requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.4
pytz==2024.1
flask==2.3.3
